refactor: move ExitCodes to the common crate
