+++ applications/tari_console_wallet/src/main.rs
@@ -22,8 +22,8 @@ use log::*;
 use opentelemetry::{self, global, KeyValue};
 use recovery::prompt_private_key_from_seed_words;
 use std::{env, process};
-use tari_app_utilities::{consts, initialization::init_configuration, utilities::ExitCodes};
-use tari_common::{configuration::bootstrap::ApplicationType, ConfigBootstrap};
+use tari_app_utilities::{consts, initialization::init_configuration};
+use tari_common::{configuration::bootstrap::ApplicationType, exit_codes::ExitCodes, ConfigBootstrap};
 use tari_common_types::types::PrivateKey;
 use tari_shutdown::Shutdown;
 use tracing_subscriber::{layer::SubscriberExt, Registry};
