+++ comms/Cargo.toml
@@ -10,6 +10,7 @@ version = "0.11.0"
 edition = "2018"
 
 [dependencies]
+tari_common = { path = "../common" }
 tari_crypto = { git = "https://github.com/tari-project/tari-crypto.git", branch = "main" }
 tari_storage = { version = "^0.11", path = "../infrastructure/storage" }
 tari_shutdown = { version = "^0.11", path = "../infrastructure/shutdown" }
