+++ comms/src/peer_manager/error.rs
@@ -21,6 +21,7 @@
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE
 
 use std::sync::PoisonError;
+use tari_common::exit_codes::ExitCodes;
 use tari_storage::KeyValStoreError;
 use thiserror::Error;
 
@@ -34,6 +35,12 @@ pub enum PeerManagerError {
     DatabaseError(#[from] KeyValStoreError),
     #[error("An error occurred while migrating the database: {0}")]
     MigrationError(String),
+}
+
+impl From<PeerManagerError> for ExitCodes {
+    fn from(err: PeerManagerError) -> Self {
+        ExitCodes::NetworkError(err.to_string())
+    }
 }
 
 impl PeerManagerError {
